+++ example/example.dart
@@ -2,7 +2,7 @@ import 'package:date_range/date_range.dart';
 
 void main(){
   final DateTime dateTime = DateTime.now();
-  final DateRange dateRange = monthDateRange(dateTime);
+  final DateRange dateRange = DateRange.month(dateTime);
   /*
   final List<data> data = db.query(lt: dateRange.lastDateTime, gt: dateRange.firstDateTime)
    */