+++ test/date_range_test.dart
@@ -4,17 +4,17 @@ import 'package:date_range/date_range.dart';
 void main() {
   test("4", (){ //april. 30 days.
     final DateTime dateTime = DateTime.utc(2019, 4, 4, 4, 4, 4, 4, 4);
-    final DateRange dateRange = monthDateRange(dateTime);
+    final DateRange dateRange = DateRange.month(dateTime);
     expect(dateRange.lastDateTime.difference(dateRange.firstDateTime).inDays, equals(30));
   });
   test("5", (){ //may. 31 days.
     final DateTime dateTime = DateTime.utc(2019, 5, 4, 4, 4, 4, 4, 4);
-    final DateRange dateRange = monthDateRange(dateTime);
+    final DateRange dateRange = DateRange.month(dateTime);
     expect(dateRange.lastDateTime.difference(dateRange.firstDateTime).inDays, equals(31));
   });
   test("12", (){ //december. 31 days. month int jumps from 12 to 1
     final DateTime dateTime = DateTime.utc(2019, 12, 4, 4, 4, 4, 4, 4);
-    final DateRange dateRange = monthDateRange(dateTime);
+    final DateRange dateRange = DateRange.month(dateTime);
     expect(dateRange.lastDateTime.difference(dateRange.firstDateTime).inDays, equals(31));
     expect(dateRange.firstDateTime.month, equals(DateTime.december));
     expect(dateRange.lastDateTime.month, equals(DateTime.january));