+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [1.0.3] - Notation changed to match Dart guidelines.
+
+DateRange's lastDateTime now represents the final microsecond of the date range instead of the first instance of the next date range.
+Please use DateRange.month(final DateTime inner) constructor.
+
 ## [1.0.2] - Added translation to text
 
 ## [1.0.1] - Now using Dart DateTime's const